some updation in departement

+++ CampusTouch.Application/Features/Departments/Commands/DeleteDepartementHandler.cs
@@ -1,8 +1,11 @@
-﻿using CampusTouch.Application.Interfaces;
+﻿using CampusTouch.Application.Common.Exceptions;
+using CampusTouch.Application.Interfaces;
 using MediatR;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Linq.Expressions;
+using System.Runtime.InteropServices;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -12,14 +15,30 @@ namespace CampusTouch.Application.Features.Departments.Commands
     {
 
         private readonly IDepartementRepository _departementRepository;
-        public DeleteDepartementHandler(IDepartementRepository DeptRepository)
+        private readonly ICurrentUserService _currentUserService;
+        public DeleteDepartementHandler(IDepartementRepository DeptRepository,ICurrentUserService currentUserService)
         {
             _departementRepository = DeptRepository;
+            _currentUserService = currentUserService;
         }
         
        public async Task<int> Handle(DeleteDepartmentCommand request,
         CancellationToken cancellationToken)
         {
+            var userid = _currentUserService.UserId;
+
+            var Existing = await _departementRepository.GetByIdAsync(request.Id);
+            if (Existing == null)
+            {
+                throw new NotFoundException("Departement Is Not Found");
+            }
+
+            if (!_currentUserService.IsAdmin)
+            {
+                throw new UnauthorizedException("only Admin can delete the departement");
+            }
+            Existing.DeletedAt = DateTime.UtcNow;
+            Existing.DeletedBy=userid;
 
             return await _departementRepository.DeleteAsync(request.Id);
         }
